% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.R
\name{compute_visium_ortho_hull}
\alias{compute_visium_ortho_hull}
\title{Compute location of segments used to create an orthogonal hull around points related
to a particular class.}
\usage{
compute_visium_ortho_hull(
  data,
  x = "x",
  y = "y",
  k = "k",
  size_x = 4.6,
  size_y = 5,
  step_x = 2.6,
  step_y = 2.4,
  delta = 0.3,
  verbose = F
)
}
\arguments{
\item{data}{a data.frame.}

\item{x}{the column name storing the x coord}

\item{y}{the column name storing the y coord}

\item{k}{the column name storing the classes (0 not part of the class of interest, 1 part of the class of interest)}

\item{size_x}{the size of the square (x axis)}

\item{size_y}{the size of the square (y axis)}

\item{step_x}{the distance between two points on the x axis.}

\item{step_y}{the distance between two points on the y axis.}

\item{delta}{add more or less flexibility to search for neighbor points}

\item{verbose}{whether the function should print (debug) info during processing.}
}
\value{
a dataframe with coordinates x, y, xend, yend (see geom_segments).
}
\description{
Compute location of segments used to create an orthogonal hull around points related
to a particular class.
}
\examples{
## Install and process the brain dataset
library(Seurat)
library(SeuratData)
library(ggplot2)
InstallData("stxBrain")
library(ohmiki)
brain <- LoadData("stxBrain", type = "anterior1")
brain <- SCTransform(brain, assay = "Spatial", verbose = FALSE)
brain <- RunPCA(brain, assay = "SCT", verbose = FALSE)
brain <- FindNeighbors(brain, reduction = "pca", dims = 1:30)
brain <- FindClusters(brain, verbose = FALSE)
brain <- RunUMAP(brain, reduction = "pca", dims = 1:30)
spatial_plot <- SpatialDimPlot(brain, label = TRUE, label.size = 3, pt.size.factor = 1.5)
## Retrieve x/y coordinates and group from ggplot object
coord_st_data <- ggplot_build(spatial_plot)$data[[1]][,c("x", "y", "group")]
coord_st_data$group <- coord_st_data$group - 1 # group are 1-based in ggplot compare to seurat
## Cluster 0 is, for instance, the cluster of interest.
cluster_to_show <- 0 # Could be also c(a, b)
coord_st_data$k <- ifelse(coord_st_data$group \%in\% cluster_to_show, 1, 0)
## Compute the segments of the hull.
path <- compute_visium_ortho_hull(coord_st_data, size_x=3.2, size_y=3.6, delta=0.5)
## Add the segments to the ggplot diagram
spatial_plot +
  theme_bw() +
  geom_segment(data=path,
               mapping=aes(x=x1,
                           y=y1,
                           xend=x2,
                           yend=y2),
               inherit.aes = F,
               color="white",
               size=0.7)
}
\keyword{hull,}
\keyword{spatial}
\keyword{transcriptomics,}
\keyword{visium}
