% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_description.R
\docType{data}
\name{coord_spot_brain}
\alias{coord_spot_brain}
\title{The spot location of the brain dataset (see stxBrain in SeuratData).}
\format{
A data frame with 2696 rows and 4 variables
}
\usage{
data(coord_spot_brain)
}
\description{
A dataset containing the x/y coordinates and class of 2696 Visium spots
}
\details{
\itemize{
\item x. The x coordinates.
\item y. The y coordinates.
\item group. A factor. The groups of the spots infered using FindClusters() (Seurat library).
}
}
\keyword{datasets}
