% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohmiki-package.R
\docType{package}
\name{ohmiki-package}
\alias{ohmiki}
\alias{ohmiki-package}
\title{ohmiki: Our helpful multi-omics tool kit}
\description{
This package implements a set of function for spatial multi-omics. Depending on our needs...
}
\author{
\strong{Maintainer}: Denis Puthier \email{denis.puthier@univ-amu.fr}

Authors:
\itemize{
  \item Arnauld Sergé \email{arnauld.serge@inserm.fr}
}

}
\keyword{internal}
