% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xy_intensity.R
\name{plot_xy_intensity}
\alias{plot_xy_intensity}
\title{Given a Seurat Spatial object, this function creates a scatter plot of the
spatial expression of a gene across spots, where the X and Y coordinates
represent the spatial location of spots and the color represents the
expression level of the gene.}
\usage{
plot_xy_intensity(
  seurat_obj = NULL,
  gene_name = NULL,
  metadata = NULL,
  intensity_slot = c("data", "counts", "sct"),
  title = "",
  legend = TRUE,
  pt_size = 2.1,
  pt_shape = 16,
  colours = c("#A9D6E5", "#2166AC", "#000000", "#B2182B", "#FFCA3A")
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing spatial expression data.}

\item{gene_name}{The name of the gene to plot.}

\item{metadata}{Provide the name of a metadata that will be used instead of genes (i.e. from
meta.data) slot of a seurat object.}

\item{intensity_slot}{The assay slot to use for the gene expression values.
Must be one of "sct", "counts", or "data". Default is "sct".}

\item{title}{The title of the plot. Default is an empty string.}

\item{legend}{Whether to display a legend for the color scale. Default is FALSE.}

\item{pt_size}{The size of the points in the plot. Default is 2.1.}

\item{pt_shape}{The shape of the points in the plot. Default is 16 (a circle).}
}
\value{
A ggplot2 object containing the scatter plot.
}
\description{
Given a Seurat Spatial object, this function creates a scatter plot of the
spatial expression of a gene across spots, where the X and Y coordinates
represent the spatial location of spots and the color represents the
expression level of the gene.
}
\examples{
library(SeuratData)
library(Seurat)
#InstallData("stxBrain")
anterior1 <- LoadData("stxBrain", type = "anterior1")
plot_xy_intensity(seurat_obj = anterior1, gene_name = "Hpca")
anterior1 <- SCTransform(anterior1, assay = "Spatial")
plot_xy_intensity(seurat_obj = anterior1, gene_name = "Hpca", intensity_slot="sct")
plot_xy_intensity(seurat_obj = anterior1, metadata = "nCount_SCT")
}
