% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_seurat_data}
\alias{preprocess_seurat_data}
\title{Preprocess scRNA-seq or spatial transcriptomics data with Seurat}
\usage{
preprocess_seurat_data(
  seurat_obj,
  normalization_method = c("NormalizeData", "SCTransform"),
  n.features = 2000,
  dims = 50,
  resolution = 0.6,
  n.neighbors = 30,
  approx = TRUE,
  seed = NULL
)
}
\arguments{
\item{seurat_obj}{A raw dataset in Seurat object format.}

\item{normalization_method}{A character string specifying the normalization method to use,
either "NormalizeData" or "SCTransform". Defaults to "NormalizeData".}

\item{n.features}{The number of features to use in the PCA analysis. Defaults to 2000.}

\item{dims}{The number of dimensions to use in the PCA analysis. Defaults to 50.}

\item{resolution}{The resolution parameter for clustering. Defaults to 0.6.}

\item{n.neighbors}{The number of neighbors for the FindNeighbors step. Defaults to 30.}

\item{approx}{Use truncated singular value decomposition to approximate PCA.}

\item{seed}{An optional seed for reproducibility.}
}
\value{
A preprocessed Seurat object.
}
\description{
This function takes a raw dataset in Seurat object format and preprocesses it using
NormalizeData or SCTransform, RunPCA, FindNeighbors, FindClusters, and RunUMAP.
}
\examples{
# Load the stxKidney dataset (spatial transcriptomics)

#InstallData('stxBrain')
LoadData("stxBrain", type = "anterior1")

# Preprocess the data
anterior1_preprocessed <- preprocess_seurat_data(anterior1, normalization_method = "SCTransform")

# Load the pbmc3k dataset (scRNA-seq)
data(pbmc3k)

# Preprocess the data
pbmc3k_preprocessed <- preprocess_seurat_data(pbmc3k)

}
