% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momicstools.R
\docType{package}
\name{scomicstools-package}
\alias{scomicstools}
\alias{scomicstools-package}
\title{scomicstools: Our helpful multi-omics tool kit}
\description{
This package implements a set of function for spatial multi-omics.
}
\keyword{internal}
